requirements.txt
flask
feedparser
schedule
sqlite3
requests
google-generativeai
python-dateutil
